refactored Order component
