+++ src/components/Order/ItemSelecter/MenuCategory/MenuCategory.css
@@ -1,0 +1,18 @@
+.mc-container
+{
+    background-color: white;
+    border: 2px solid black;
+    width: 6em;
+    height: 4em;
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    overflow: hidden;
+}
+.mc-container > *
+{
+    text-align: center;
+    margin: 0;
+    font-size: 1.4em;
+}
+
