+++ src/components/Order/Order.css
@@ -1,0 +1,6 @@
+.order-container
+{
+    display: flex;
+    flex-wrap: wrap;
+    justify-content: center;
+}
