+++ src/components/Order/Order.js
@@ -1,7 +1,8 @@
 import React from "react";
-import ItemSelecter from "./ItemSelecter";
-import ItemSummary from "./ItemSummary";
+import ItemSelecter from "./ItemSelecter/ItemSelecter";
+import ItemSummary from "./ItemSummary/ItemSummary";
 import {Spring} from 'react-spring/renderprops';
+import "./Order.css";
 
 
 class Order extends React.Component
@@ -38,7 +39,7 @@ class Order extends React.Component
         to={{opacity: 1, marginTop: 0}}
       >
         {props => (
-          <div style={Object.assign( {}, container, props)} >
+          <div style={Object.assign( {}, props)} className="order-container" >
             <ItemSelecter
             data={ {selectedItem: this.state.selectedItem, 
                     updateSelectedItem: this.getSelectedItem }}
